Fix MSVC warnings about implicit value truncations.
Cache sizes are required to fit in an unsigned char, so make this explicit.
This prevents MSVC from warning in various places that we may be truncating
an int to fit in an unsigned char.

This means we need to move the enforcement tests to
VCDiffStreamingDecoderImpl::InitCustomCodeTable(),
and, VCDiffAddressCache::Init() can no longer fail.
This also results in moving some unittest content around.