vcdecoder: add memory-based output interface
Currently, the VCDiffStreamingDecoder class expects applications to call
into ::StartDecoding() with a dictionary, and then maintains an internal
buffer (decoded_target_) to accumulate pending output bytes. These are then
written through ::AppendNewOutputText() and ::FlushDecodedTarget().

The problem with this implementation is that the allocation of this
intermediate buffer may fail, which will then result in the application
abort with a trap.

This patch adds an alternative interface that does not lead to internal
allocation. StartDecoding() may now optionally be called with an output
buffer and a maximum length, and DecodeChunk() may be called without an
OutputString parameter. If operated this way, the decoder will directly
write to the provided memory buffer, which can, for instance, be mmap()ed
on a block device.

The old interface is kept around for full compatibility.